requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0
pillow>=9.0.0
git+https://github.com/Kaggle/kagglehub.git
kagglehub>=0.0.12
tensorboard
